requirements.txt
Flask==3.0.0
Flask-RESTX==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
PyMySQL==1.1.0
requests==2.31.0
httpx==0.25.2
authlib==1.2.1
python-jose==3.3.0
cryptography==41.0.7
python-dotenv==1.0.0
marshmallow==3.20.1
fhir.resources==7.0.2
python-dateutil==2.8.2
pandas==2.1.4
elasticsearch==8.11.0
gunicorn==21.2.0
PyJWT==2.8.0
